# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JetKöy API specification
 * OpenAPI spec version: 0.1.0
 */
export interface SuccessResponse {
  success: boolean;
}

export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface RegisterRequest {
  fullName: string;
  phone: string;
  plate: string;
  password: string;
}

export interface LoginRequest {
  phone: string;
  password: string;
}

export interface User {
  id: string;
  fullName: string;
  phone: string;
  plate: string;
  credits: number;
  isVip: boolean;
  isAdmin: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  available: 'available',
  grabbed: 'grabbed',
} as const;

export interface Job {
  id: string;
  creatorId: string;
  creatorName: string;
  creatorPlate: string;
  departure: string;
  destination: string;
  passengerName: string;
  passengerPhoneMasked: string;
  totalFare: number;
  commission: number;
  status: JobStatus;
  createdAt: string;
}

export type JobDetailStatus = typeof JobDetailStatus[keyof typeof JobDetailStatus];


export const JobDetailStatus = {
  available: 'available',
  grabbed: 'grabbed',
} as const;

export interface JobDetail {
  id: string;
  creatorId: string;
  creatorName: string;
  creatorPlate: string;
  departure: string;
  destination: string;
  passengerName: string;
  passengerPhone: string;
  totalFare: number;
  commission: number;
  status: JobDetailStatus;
  grabbedById?: string | null;
  createdAt: string;
}

export interface GrabJobResponse {
  job: JobDetail;
  passengerPhone: string;
  newCreditBalance: number;
}

export interface CreateJobRequest {
  departure: string;
  destination: string;
  passengerName: string;
  passengerPhone: string;
  totalFare: number;
  commission: number;
}

export interface MyJobsSummary {
  createdJobs: Job[];
  grabbedJobs: Job[];
}

export interface Wallet {
  credits: number;
  totalEarned: number;
  totalSpent: number;
}

export type CommissionRecordStatus = typeof CommissionRecordStatus[keyof typeof CommissionRecordStatus];


export const CommissionRecordStatus = {
  pending: 'pending',
  paid: 'paid',
} as const;

export interface CommissionRecord {
  id: string;
  jobId: string;
  departure: string;
  destination: string;
  amount: number;
  driverName: string;
  driverPlate: string;
  status: CommissionRecordStatus;
  createdAt: string;
}

export interface CommissionSummary {
  receivable: CommissionRecord[];
  payable: CommissionRecord[];
}

export interface AdminUser {
  id: string;
  fullName: string;
  phone: string;
  plate: string;
  credits: number;
  isVip: boolean;
  isAdmin: boolean;
  jobsCreated: number;
  jobsGrabbed: number;
  createdAt: string;
}

export interface AdminEditUserRequest {
  fullName?: string;
  phone?: string;
  plate?: string;
  credits?: number;
  isVip?: boolean;
  isAdmin?: boolean;
}

export interface UpdateProfileRequest {
  fullName?: string;
  plate?: string;
  currentPassword?: string;
  newPassword?: string;
}

export interface AddCreditsRequest {
  amount: number;
  note?: string;
}

export interface SetVipRequest {
  isVip: boolean;
}

export interface SendMessageRequest {
  content: string;
}

export interface UserMessage {
  id: string;
  content: string;
  isRead: boolean;
  adminReply?: string | null;
  repliedAt?: string | null;
  createdAt: string;
}

export interface AdminMessage {
  id: string;
  senderId: string;
  senderName: string;
  senderPhone: string;
  senderPlate: string;
  content: string;
  isRead: boolean;
  adminKeep: boolean;
  adminReply?: string | null;
  repliedAt?: string | null;
  createdAt: string;
}

export interface AdminReplyRequest {
  reply: string;
}

export interface AdminSendMessageRequest {
  userId: string;
  content: string;
}

export interface SendOtpRequest {
  phone: string;
}

export interface SendOtpResponse {
  message: string;
  devCode?: string | null;
}

export interface VerifyOtpRequest {
  fullName: string;
  phone: string;
  plate: string;
  password: string;
  code: string;
}

export type UpdatePushTokenBody = {
  pushToken: string;
};